Fix regeneration of assemble images if input transform changed Fixed sorting of crop box resulting in incorrect volume boundaries Fixed converting string list of integers to single integer in assemble pipleine Fixes to .cmd files

+++ nornir_buildmanager/operations/tile.py
@@ -940,9 +940,11 @@ def AssembleTransformScipy(Parameters, Logger, FilterNode, TransformNode, Output
 
     if hasattr(ImageNode, 'InputTransformChecksum'):
         if not transforms.IsValueMatched(ImageNode, 'InputTransformChecksum', TransformNode.Checksum):
-            os.remove(ImageNode.FullPath)
+            if os.path.exists(ImageNode.FullPath):
+                os.remove(ImageNode.FullPath)
     else:
-        os.remove(ImageNode.FullPath)
+        if os.path.exists(ImageNode.FullPath):
+            os.remove(ImageNode.FullPath)
 
 
     if not (os.path.exists(ImageNode.FullPath) and os.path.exists(MaskImageNode.FullPath)):
@@ -990,6 +992,9 @@ def AssembleTransformScipy(Parameters, Logger, FilterNode, TransformNode, Output
 
         shutil.move(tempOutputFullPath, ImageNode.FullPath)
         shutil.move(tempMaskOutputFullPath, MaskImageNode.FullPath)
+
+        ImageNode.InputTransformChecksum = TransformNode.Checksum
+        MaskImageNode.InputTransformChecksum = TransformNode.Checksum
 
         # ImageNode.Checksum = nornir_shared.Checksum.FilesizeChecksum(ImageNode.FullPath)
         # MaskImageNode.Checksum = nornir_shared.Checksum.FilesizeChecksum(MaskImageNode.FullPath)
